pyproject.toml
[project]
name = "meet-mind"
version = "0.1.0"
description = "On-device meeting agent prototype using one Gemma model with MLX."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mlx-vlm>=0.1.10",
  "faster-whisper>=1.1.1",
  "numpy>=1.26.0",
  "opencv-python>=4.10.0.84",
  "pillow>=10.4.0",
  "sounddevice>=0.5.1",
]

[project.scripts]
roomscribe = "backend.agents.roomscribe.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
